class02-20250420/04_agent_hello_world/agent_hello/pyproject.toml
[project]
name = "agent-hello"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.11",
    "python-dotenv>=1.1.0",
]

[project.scripts]
agent-hello = "agent_hello:main"
global = "agent_hello.global_config:main"
runner = "agent_hello.runner_config:main"
agent = "agent_hello.agent_config:main"
async_run = "agent_hello.async_run:run_main"
with_tool = "agent_hello.with_tool:main"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
